backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-docx==1.1.2
reportlab==4.4.10
httpx==0.27.2
python-multipart==0.0.9
python-dotenv==1.0.1
